refactored script into new repos
