+++ src/main/java/com/tidal/wave/options/EdgeBrowserOptions.java
@@ -1,6 +1,6 @@
 package com.tidal.wave.options;
 
-import com.tidal.wave.propertieshandler.Config;
+import com.tidal.wave.config.Config;
 import org.openqa.selenium.edge.EdgeOptions;
 
 import java.util.HashMap;
