--- src/main/java/com/tidal/wave/report/ReportMatcher.java
@@ -1,11 +1,0 @@
-package com.tidal.wave.report;
-
-public abstract class ReportMatcher {
-    protected ReportMatcher nextReportTypeMatcher;
-
-    public void setNextMatcher(ReportMatcher nextReportTypeMatcher){
-        this.nextReportTypeMatcher = nextReportTypeMatcher;
-    }
-
-    public abstract ReportModel parse(String typeContent, String errorMessage);
-}
