--- src/main/java/com/tidal/wave/utils/CheckString.java
@@ -1,16 +1,0 @@
-package com.tidal.wave.utils;
-
-public class CheckString {
-
-    //Not to be instantiated
-    private CheckString() {
-    }
-
-    public static boolean isNullOrEmpty(String value){
-        return value == null || value.isEmpty();
-    }
-
-    public static boolean isNotNullOrEmpty(String value){
-        return value != null && !value.isEmpty();
-    }
-}
