+++ src/test/java/com/tidal/wave/browser/OpenTest.java
@@ -1,6 +1,5 @@
 package com.tidal.wave.browser;
 
-import com.tidal.wave.loggers.Logger;
 import com.tidal.wave.page.Page;
 import org.junit.Before;
 import org.junit.Test;
@@ -8,13 +7,15 @@ import org.openqa.selenium.chrome.ChromeOptions;
 import org.openqa.selenium.edge.EdgeOptions;
 import org.openqa.selenium.firefox.FirefoxOptions;
 import org.openqa.selenium.remote.RemoteWebDriver;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.*;
 
 public class OpenTest {
 
-    Logger logger = new Logger(OpenTest.class);
+    Logger logger = LoggerFactory.getLogger(OpenTest.class);
 
     @Before
     public void init(){
